rename `config-manager.js` to `config.js`
