+++ packages/utils/index.js
@@ -2,5 +2,5 @@ export { default as onExit } from './lib/on-exit.js';
 export { default as execScript } from './lib/exec-script.js';
 export { default as launchFromCli } from './lib/launch-from-cli.js';
 export { default as LogManager, Logger } from './lib/log-manager.js';
-export * from './lib/config-manager.js';
+export * from './lib/config.js';
 export { default as TaskQueue, Task } from './lib/task-queue.js';
